+++ src/main/java/com/clearblade/java/api/MessageCallback.java
@@ -2,24 +2,25 @@ package com.clearblade.java.api;
 
 public class MessageCallback {
 	public void done(String topic, String message){
-		//override to get a string message
+		// override to get a string message
 	}
+
 	public void error(ClearBladeException exception){
-		//override to catch an error
+		// override to catch an error
 	}
 	
 	public void done(String topic, Item message){
-		//override to get an Item message
+		// override to get an Item message
 	}
 	
 	public void done(String topic, byte[] message){
-		//override to get a byte[] message
+		// override to get a byte[] message
 	}
 	public void done(String topic, String message, int qos) {
-		
+	    // override to get a string message with qos
 	}
 	
 	public void done(History history){
-		
+	    // override to get History
 	}
 }
